feat(testcase): implement permission checks and refactor serial type in testcase operations

+++ internal/app/dto/response/testcase.go
@@ -3,7 +3,7 @@ package response
 type TestcaseData struct {
 	Id         uint64 `json:"id,omitempty"`
 	ProblemId  uint64 `json:"problem_id"`
-	Serial     uint64 `json:"serial"`
+	Serial     uint16 `json:"serial"`
 	TestInput  string `json:"test_input"`
 	TestOutput string `json:"test_output"`
 }
